+++ ebs-20210730/alibabacloud_ebs20210730/models.py
@@ -7174,44 +7174,66 @@ class DescribeMetricDataRequest(TeaModel):
         aggre_over_line_ops: str = None,
         dimensions: str = None,
         end_time: str = None,
+        group_by_labels: List[str] = None,
         metric_name: str = None,
         period: int = None,
         region_id: str = None,
         start_time: str = None,
     ):
-        # Aggregation method in time dimension. Valid values:
+        # Aggregation method over time. Possible values include:
         # 
-        # - SUM
-        # - COUNT
-        # - AVG
-        # - MAX
-        # - MIN
+        # - SUM_OVER_TIME
+        # - COUNT_OVER_TIME
+        # - AVG_OVER_TIME
+        # - MAX_OVER_TIME
+        # - MIN_OVER_TIME
+        # - SUM_OVER_TIME_LCRO: Sum over a left-closed, right-open interval
+        # - AVG_OVER_TIME_LCRO: Average over a left-closed, right-open interval
+        # - SUM_OVER_TIME_LORC: Sum over a left-open, right-closed interval
+        # - AVG_OVER_TIME_LORC: Average over a left-open, right-closed interval
         self.aggre_ops = aggre_ops
+        # Aggregation method between lines. Possible values include:
+        # - NON: No aggregation
+        # - SUM: Sum
+        # - AVG: Average
+        # - COUNT: Count
+        # - MAX: Maximum
+        # - MIN: Minimum
         self.aggre_over_line_ops = aggre_over_line_ops
-        # The dimension map in the JSON format. A dimension is a key-value pair. Valid dimension key: diskId.
-        self.dimensions = dimensions
-        # The end of the time range to query. The specified time must be later than the current time. Specify the time in the ISO 8601 standard in the yyyy-MM-ddTHH:mm:ssZ format. The time must be in UTC.
-        self.end_time = end_time
-        # The name of the metric. Valid values:
+        # Dimension map, in JSON format, representing the dimensions being queried. The currently available keys are:
+        # - DiskId: Cloud disk name, e.g., d-xxx.
+        # - DeviceType: Type of cloud disk, system indicates system disk, data indicates data disk.
+        # - DeviceCategory: Category of cloud disk, e.g., cloud_essd.
+        # - EcsInstanceId: Name of the ECS instance where the disk is located, e.g., i-xxx.
         # 
-        # *   disk_bps_percent
-        # *   disk_iops_percent
-        # *   disk_read_block_size
-        # *   disk_read_bps
-        # *   disk_read_iops
-        # *   disk_read_latency
-        # *   disk_write_block_size
-        # *   disk_write_bps
-        # *   disk_write_iops
-        # *   disk_write_latency
+        # The returned results are the intersection of all dimension filter conditions.
+        self.dimensions = dimensions
+        # The end time point for obtaining metric data. It should not be later than the current moment. Represented according to the ISO 8601 standard, using UTC +0 time, in the format yyyy-MM-ddTHH:mm:ssZ.
+        self.end_time = end_time
+        self.group_by_labels = group_by_labels
+        # Metric name. Possible values include:
+        #    
+        # - disk_bps_percent
+        # - disk_iops_percent
+        # - disk_read_block_size
+        # - disk_read_bps
+        # - disk_read_iops
+        # - disk_write_block_size
+        # - disk_write_bps
+        # - disk_write_iops
         # 
         # This parameter is required.
         self.metric_name = metric_name
-        # The interval at which metric data is collected. Unit: seconds. Default value: 60. Valid values: 60, 300, 600, and 3600, which support queries for time ranges of up to 2 hours, 2 hours, 1 day, and 7 days, respectively. For example, if you set Period to 60, the end time is less than 2 hours from the end time.
+        # The interval for obtaining metric data. Unit: seconds. The default value is 5 seconds. Possible values include:
+        # 
+        # - 5: 5s precision query, can query up to 12 hours of data
+        # - 10: 10s precision query, can query up to 24 hours of data
+        # - 60: 60s precision query, can query up to 7 days of data
+        # - 3600: 3600s precision query, can query up to 30 days of data
         self.period = period
-        # The region ID.
+        # Region ID.
         self.region_id = region_id
-        # The beginning of the time range to query. You can specify a point in time that is up to one year apart from the current time. If StartTime and EndTime are both unspecified, the monitoring metric data of the last Period value is queried. Specify the time in the ISO 8601 standard in the yyyy-MM-ddTHH:mm:ssZ format. The time must be in UTC.
+        # The start time point for obtaining metric data. The earliest selectable time is one year before the current moment. When both StartTime and EndTime parameters are empty, it defaults to querying the most recent period\\"s monitoring metrics. Represented according to the ISO 8601 standard, using UTC +0 time, in the format yyyy-MM-ddTHH:mm:ssZ.
         self.start_time = start_time
 
     def validate(self):
@@ -7231,6 +7253,8 @@ class DescribeMetricDataRequest(TeaModel):
             result['Dimensions'] = self.dimensions
         if self.end_time is not None:
             result['EndTime'] = self.end_time
+        if self.group_by_labels is not None:
+            result['GroupByLabels'] = self.group_by_labels
         if self.metric_name is not None:
             result['MetricName'] = self.metric_name
         if self.period is not None:
@@ -7251,6 +7275,129 @@ class DescribeMetricDataRequest(TeaModel):
             self.dimensions = m.get('Dimensions')
         if m.get('EndTime') is not None:
             self.end_time = m.get('EndTime')
+        if m.get('GroupByLabels') is not None:
+            self.group_by_labels = m.get('GroupByLabels')
+        if m.get('MetricName') is not None:
+            self.metric_name = m.get('MetricName')
+        if m.get('Period') is not None:
+            self.period = m.get('Period')
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
+        return self
+
+
+class DescribeMetricDataShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        aggre_ops: str = None,
+        aggre_over_line_ops: str = None,
+        dimensions: str = None,
+        end_time: str = None,
+        group_by_labels_shrink: str = None,
+        metric_name: str = None,
+        period: int = None,
+        region_id: str = None,
+        start_time: str = None,
+    ):
+        # Aggregation method over time. Possible values include:
+        # 
+        # - SUM_OVER_TIME
+        # - COUNT_OVER_TIME
+        # - AVG_OVER_TIME
+        # - MAX_OVER_TIME
+        # - MIN_OVER_TIME
+        # - SUM_OVER_TIME_LCRO: Sum over a left-closed, right-open interval
+        # - AVG_OVER_TIME_LCRO: Average over a left-closed, right-open interval
+        # - SUM_OVER_TIME_LORC: Sum over a left-open, right-closed interval
+        # - AVG_OVER_TIME_LORC: Average over a left-open, right-closed interval
+        self.aggre_ops = aggre_ops
+        # Aggregation method between lines. Possible values include:
+        # - NON: No aggregation
+        # - SUM: Sum
+        # - AVG: Average
+        # - COUNT: Count
+        # - MAX: Maximum
+        # - MIN: Minimum
+        self.aggre_over_line_ops = aggre_over_line_ops
+        # Dimension map, in JSON format, representing the dimensions being queried. The currently available keys are:
+        # - DiskId: Cloud disk name, e.g., d-xxx.
+        # - DeviceType: Type of cloud disk, system indicates system disk, data indicates data disk.
+        # - DeviceCategory: Category of cloud disk, e.g., cloud_essd.
+        # - EcsInstanceId: Name of the ECS instance where the disk is located, e.g., i-xxx.
+        # 
+        # The returned results are the intersection of all dimension filter conditions.
+        self.dimensions = dimensions
+        # The end time point for obtaining metric data. It should not be later than the current moment. Represented according to the ISO 8601 standard, using UTC +0 time, in the format yyyy-MM-ddTHH:mm:ssZ.
+        self.end_time = end_time
+        self.group_by_labels_shrink = group_by_labels_shrink
+        # Metric name. Possible values include:
+        #    
+        # - disk_bps_percent
+        # - disk_iops_percent
+        # - disk_read_block_size
+        # - disk_read_bps
+        # - disk_read_iops
+        # - disk_write_block_size
+        # - disk_write_bps
+        # - disk_write_iops
+        # 
+        # This parameter is required.
+        self.metric_name = metric_name
+        # The interval for obtaining metric data. Unit: seconds. The default value is 5 seconds. Possible values include:
+        # 
+        # - 5: 5s precision query, can query up to 12 hours of data
+        # - 10: 10s precision query, can query up to 24 hours of data
+        # - 60: 60s precision query, can query up to 7 days of data
+        # - 3600: 3600s precision query, can query up to 30 days of data
+        self.period = period
+        # Region ID.
+        self.region_id = region_id
+        # The start time point for obtaining metric data. The earliest selectable time is one year before the current moment. When both StartTime and EndTime parameters are empty, it defaults to querying the most recent period\\"s monitoring metrics. Represented according to the ISO 8601 standard, using UTC +0 time, in the format yyyy-MM-ddTHH:mm:ssZ.
+        self.start_time = start_time
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.aggre_ops is not None:
+            result['AggreOps'] = self.aggre_ops
+        if self.aggre_over_line_ops is not None:
+            result['AggreOverLineOps'] = self.aggre_over_line_ops
+        if self.dimensions is not None:
+            result['Dimensions'] = self.dimensions
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+        if self.group_by_labels_shrink is not None:
+            result['GroupByLabels'] = self.group_by_labels_shrink
+        if self.metric_name is not None:
+            result['MetricName'] = self.metric_name
+        if self.period is not None:
+            result['Period'] = self.period
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AggreOps') is not None:
+            self.aggre_ops = m.get('AggreOps')
+        if m.get('AggreOverLineOps') is not None:
+            self.aggre_over_line_ops = m.get('AggreOverLineOps')
+        if m.get('Dimensions') is not None:
+            self.dimensions = m.get('Dimensions')
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+        if m.get('GroupByLabels') is not None:
+            self.group_by_labels_shrink = m.get('GroupByLabels')
         if m.get('MetricName') is not None:
             self.metric_name = m.get('MetricName')
         if m.get('Period') is not None:
@@ -7268,9 +7415,9 @@ class DescribeMetricDataResponseBodyDataList(TeaModel):
         datapoints: Any = None,
         labels: Any = None,
     ):
-        # The datapoints that consist of consecutive timestamps in seconds and metric values that were recorded at these time points.
+        # List of monitoring data, consisting of a series of consecutive second-level timestamps and the corresponding metric values at those times.
         self.datapoints = datapoints
-        # The tags.
+        # Labels.
         self.labels = labels
 
     def validate(self):
@@ -7305,12 +7452,13 @@ class DescribeMetricDataResponseBody(TeaModel):
         total_count: int = None,
         warnings: List[str] = None,
     ):
-        # The disk monitoring data.
+        # Collection of monitoring data for the cloud disk.
         self.data_list = data_list
-        # The request ID.
+        # Request ID.
         self.request_id = request_id
-        # The total number of data entries queried.
+        # Total number of data points queried.
         self.total_count = total_count
+        # List of warning messages.
         self.warnings = warnings
 
     def validate(self):
@@ -9161,7 +9309,6 @@ class ListReportsRequest(TeaModel):
     def __init__(
         self,
         app_id: str = None,
-        app_name: str = None,
         max_results: int = None,
         next_token: str = None,
         page_number: int = None,
@@ -9169,8 +9316,6 @@ class ListReportsRequest(TeaModel):
         region_id: str = None,
     ):
         self.app_id = app_id
-        # App name.
-        self.app_name = app_name
         # Maximum number of items for Token-based pagination.
         self.max_results = max_results
         # Query token (Token), the value is the NextToken parameter value returned from the previous API call.
@@ -9193,8 +9338,6 @@ class ListReportsRequest(TeaModel):
         result = dict()
         if self.app_id is not None:
             result['AppId'] = self.app_id
-        if self.app_name is not None:
-            result['AppName'] = self.app_name
         if self.max_results is not None:
             result['MaxResults'] = self.max_results
         if self.next_token is not None:
@@ -9211,8 +9354,6 @@ class ListReportsRequest(TeaModel):
         m = m or dict()
         if m.get('AppId') is not None:
             self.app_id = m.get('AppId')
-        if m.get('AppName') is not None:
-            self.app_name = m.get('AppName')
         if m.get('MaxResults') is not None:
             self.max_results = m.get('MaxResults')
         if m.get('NextToken') is not None: